% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hypotenuse_length.R
\name{get_hypotenuse_length}
\alias{get_hypotenuse_length}
\title{Calculate the length of the hypotenuse of a right triangle}
\usage{
get_hypotenuse_length(height, base)
}
\arguments{
\item{height}{a numeric value of the height of a right triangle}

\item{base}{a numeric value of the base of a right triangle}
}
\value{
The length of the hypotenuse of the right triangle
}
\description{
Calculate the length of the hypotenuse of a right triangle
}
