% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimmed_mean.R
\name{trimmed_mean}
\alias{trimmed_mean}
\title{Calculate a trimmed mean for a numeric vector}
\usage{
trimmed_mean(x, l, s)
}
\arguments{
\item{x}{a numeric vector}

\item{l}{the l largest values}

\item{s}{the s smallerst values}
}
\description{
Calculate a trimmed mean for a numeric vector
}
